refactor(convert): move inline_root.rs tests into home module
